+++ src/anemoi/inference/grib/templates/file.py
@@ -8,46 +8,81 @@
 # nor does it submit to any jurisdiction.
 
 import logging
+from functools import cached_property
+from pathlib import Path
 from typing import Any
+from typing import Literal
 
 import earthkit.data as ekd
 
+from anemoi.inference.decorators import main_argument
+from anemoi.inference.inputs.ekd import find_variable
+from anemoi.inference.types import State
+
 from . import TemplateProvider
 from . import template_provider_registry
+from .manager import TemplateManager
 
 LOG = logging.getLogger(__name__)
 
 
 @template_provider_registry.register("file")
+@main_argument("path")
 class FileTemplates(TemplateProvider):
     """Template provider using a single GRIB file."""
 
-    def __init__(self, manager: Any, path: str) -> None:
+    def __init__(
+        self,
+        manager: TemplateManager,
+        *,
+        path: str,
+        mode: Literal["auto", "first", "last"] = "first",
+        variables: str | list | None = None,
+    ) -> None:
         """Initialize the FileTemplates instance.
 
         Parameters
         ----------
-        manager : Any
+        manager : TemplateManager
             The manager instance.
         path : str
             The path to the GRIB file.
+        mode : Literal["auto", "first", "last"], optional
+            The method with which to select a message from the grib file to use as template, by default "first":
+            - "first": use the first message in the grib file
+            - "last": use the last message in the grib file
+            - "auto": select variable from the grib file matching the output variable name
+        variables : str | list, optional
+            The output variable name(s) for which to use this template file. If empty, applies to all variables.
         """
         self.manager = manager
-        self.path = path
+        self.path = Path(path)
+        if not self.path.exists():
+            raise FileNotFoundError(f"GRIB template file not found: {self.path}")
+        self.mode = mode
+        self.variables = variables if isinstance(variables, list) else [variables] if variables else None
 
-    def template(self, grib: str, lookup: dict[str, Any]) -> ekd.Field:
-        """Retrieve the template from the GRIB file.
+    def __repr__(self):
+        info = f"{self.__class__.__name__}({self.path.name},mode={self.mode}{{variables}})"
+        return info.format(variables=f",variables={self.variables}" if self.variables else "")
 
-        Parameters
-        ----------
-        grib : str
-            The GRIB string.
-        lookup : Dict[str, Any]
-            The lookup dictionary.
+    @cached_property
+    def _data(self):
+        return ekd.from_source("file", self.path)
 
-        Returns
-        -------
-        ekd.Field
-            The field from the GRIB file.
-        """
-        return ekd.from_source("file", self.path)[0]
+    def template(self, variable: str, lookup: dict[str, Any], state: State, **kwargs) -> ekd.Field | None:
+        if self.variables and variable not in self.variables:
+            return None
+
+        match self.mode:
+            case "first":
+                return self._data[0]
+            case "last":
+                return self._data[-1]
+            case "auto":
+                namer = getattr(state.get("_input"), "_namer", self.manager.owner.context.checkpoint.default_namer())
+                field = find_variable(self._data, variable, namer)
+                if len(field) > 0:
+                    return field[0]
+
+        return None